requirements.txt
numpy==1.26.4
keyboard==0.13.5
pillow==10.3.0
google-cloud-storage==2.16.0
matplotlib==3.9.0
pywebview==5.4
websockets==15.0.1
